refactored out helper functions
