+++ src/crweb/tests.py
@@ -1,3 +1,9 @@
+######################################################################
+# Runbook Web Application
+# -------------------------------------------------------------------
+# Tests
+######################################################################
+
 import unittest
 import rethinkdb as r
 from rethinkdb.errors import RqlDriverError
